gas-ml/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Machine Learning
torch==2.1.2
xgboost==2.0.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Blockchain
web3==6.13.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
# Utilities
tqdm==4.66.1

# Benchmarking
statsmodels==0.14.0
prophet==1.1.5
